Fixing warnings, speeding up compilation

+++ praticals/3_Integration/rk4.cpp
@@ -9,7 +9,7 @@ struct Derivative {
   dvec3 dx, dv;
 };
 
-dvec3 acceleration(const sBall &body, dvec3 x, dvec3 v, float t) {
+dvec3 acceleration(const sBall &body, dvec3 x, dvec3 v, double t) {
   // we could be summing accelerations, or doing other cool things here
   return dvec3(0, -10.0, 0);
 }
